+++ authenticator/src/main/java/com/amplifyframework/ui/authenticator/ui/PasswordInputField.kt
@@ -26,6 +26,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
@@ -35,6 +36,7 @@ import com.amplifyframework.ui.authenticator.R
 import com.amplifyframework.ui.authenticator.forms.FieldConfig
 import com.amplifyframework.ui.authenticator.forms.MutablePasswordFieldState
 import com.amplifyframework.ui.authenticator.strings.StringResolver
+import com.amplifyframework.ui.authenticator.util.contentTypeForKey
 
 @Composable
 internal fun PasswordInputField(
@@ -54,7 +56,7 @@ internal fun PasswordInputField(
     )
 
     OutlinedTextField(
-        modifier = modifier,
+        modifier = modifier.contentTypeForKey(fieldConfig.key),
         enabled = enabled,
         value = fieldState.content,
         onValueChange = { fieldState.content = it },
@@ -91,7 +93,7 @@ private fun getTrailingIcon(visible: Boolean, onClick: () -> Unit): @Composable 
             true -> R.string.amplify_ui_authenticator_field_a11y_password_hide
             false -> R.string.amplify_ui_authenticator_field_a11y_password_show
         }
-        IconButton(onClick = onClick) {
+        IconButton(onClick = onClick, modifier = Modifier.testTag(TestTags.ShowPasswordIcon)) {
             Icon(painter = painterResource(icon), contentDescription = stringResource(contentDescription))
         }
     }