+++ authenticator/src/main/java/com/amplifyframework/ui/authenticator/ui/PasswordResetConfirm.kt
@@ -24,6 +24,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalAutofillManager
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.amplifyframework.auth.AuthCodeDeliveryDetails
@@ -49,7 +51,7 @@ fun PasswordResetConfirm(
             .padding(horizontal = 16.dp)
     ) {
         headerContent(state)
-        deliveryNoticeContent(details = state.deliveryDetails)
+        deliveryNoticeContent(state.deliveryDetails)
         AuthenticatorForm(
             state = state.form
         )
@@ -62,12 +64,18 @@ fun PasswordResetConfirm(
 }
 
 @Composable
-fun PasswordResetConfirmFooter(
-    state: PasswordResetConfirmState,
-    modifier: Modifier = Modifier
-) {
+fun PasswordResetConfirmFooter(state: PasswordResetConfirmState, modifier: Modifier = Modifier) {
+    // If we navigate away from the screen via some reason other that submitting the form then cancel any changes
+    // in autofill manager so that user won't be prompted to update their saved password
+    val autofillManager = LocalAutofillManager.current
     Column(modifier = modifier.fillMaxWidth(), horizontalAlignment = Alignment.CenterHorizontally) {
-        TextButton(onClick = { state.moveTo(AuthenticatorStep.SignIn) }) {
+        TextButton(
+            modifier = Modifier.testTag(TestTags.BackToSignInButton),
+            onClick = {
+                autofillManager?.cancel()
+                state.moveTo(AuthenticatorStep.SignIn)
+            }
+        ) {
             Text(stringResource(R.string.amplify_ui_authenticator_button_back_to_signin))
         }
     }