+++ authenticator/src/test/java/com/amplifyframework/ui/authenticator/ui/SignInTest.kt
@@ -15,13 +15,21 @@
 
 package com.amplifyframework.ui.authenticator.ui
 
-import com.amplifyframework.ui.authenticator.auth.SignInMethod
-import com.amplifyframework.ui.authenticator.states.SignInStateImpl
+import androidx.compose.ui.autofill.AutofillManager
+import androidx.compose.ui.autofill.ContentType
+import com.amplifyframework.ui.authenticator.enums.AuthenticatorStep
+import com.amplifyframework.ui.authenticator.forms.FieldError
+import com.amplifyframework.ui.authenticator.forms.FieldKey
+import com.amplifyframework.ui.authenticator.forms.setFieldError
+import com.amplifyframework.ui.authenticator.testUtil.AuthenticatorUiTest
+import com.amplifyframework.ui.authenticator.testUtil.mockSignInState
 import com.amplifyframework.ui.authenticator.ui.robots.signIn
-import com.amplifyframework.ui.testing.ComposeTest
+import com.amplifyframework.ui.testing.ScreenshotTest
+import io.mockk.mockk
+import io.mockk.verify
 import org.junit.Test
 
-class SignInTest : ComposeTest() {
+class SignInTest : AuthenticatorUiTest() {
 
     @Test
     fun `title is Sign In`() {
@@ -43,9 +51,92 @@ class SignInTest : ComposeTest() {
         }
     }
 
-    private fun mockSignInState() = SignInStateImpl(
-        signInMethod = SignInMethod.Username,
-        onSubmit = { _, _ -> },
-        onMoveTo = { }
-    )
+    @Test
+    fun `has expected content types set`() {
+        setContent(providedStep = AuthenticatorStep.SignIn) {
+            SignIn(state = mockSignInState())
+        }
+        signIn {
+            hasUsernameContentType(ContentType.Username)
+            hasPasswordContentType(ContentType.Password)
+        }
+    }
+
+    @Test
+    fun `cancels autofill values on create account`() {
+        val autofillManager = mockk<AutofillManager>(relaxed = true)
+        setContent(autofillManager = autofillManager) {
+            SignIn(state = mockSignInState())
+        }
+        signIn {
+            setUsername("foo")
+            setPassword("bar")
+            clickCreateAccount()
+        }
+        verify {
+            autofillManager.cancel()
+        }
+    }
+
+    @Test
+    fun `cancels autofill values on forgot password`() {
+        val autofillManager = mockk<AutofillManager>(relaxed = true)
+        setContent(autofillManager = autofillManager) {
+            SignIn(state = mockSignInState())
+        }
+        signIn {
+            setUsername("foo")
+            setPassword("bar")
+            clickForgotPassword()
+        }
+        verify {
+            autofillManager.cancel()
+        }
+    }
+
+    @Test
+    @ScreenshotTest
+    fun `default state`() {
+        setContent {
+            SignIn(state = mockSignInState())
+        }
+    }
+
+    @Test
+    @ScreenshotTest
+    fun `ready to submit`() {
+        setContent {
+            SignIn(state = mockSignInState())
+        }
+        signIn {
+            setUsername("username")
+            setPassword("password")
+        }
+    }
+
+    @Test
+    @ScreenshotTest
+    fun `password visible`() {
+        setContent {
+            SignIn(state = mockSignInState())
+        }
+        signIn {
+            setUsername("username")
+            setPassword("password")
+            clickShowPassword()
+        }
+    }
+
+    @Test
+    @ScreenshotTest
+    fun `username not found`() {
+        val state = mockSignInState()
+        setContent {
+            SignIn(state = state)
+        }
+        signIn {
+            setUsername("username")
+        }
+        state.form.setFieldError(FieldKey.Username, FieldError.NotFound)
+    }
 }