+++ build-logic/plugins/src/main/kotlin/ComponentConventionPlugin.kt
@@ -34,10 +34,11 @@ class ComponentConventionPlugin : Plugin<Project> {
             pluginManager.apply("amplify.android.kover")
             pluginManager.apply("amplify.android.api.validator")
             pluginManager.apply("amplify.android.licenses")
+            pluginManager.apply("amplify.android.screenshots")
 
             tasks.withType<KotlinCompile>().configureEach {
-                kotlinOptions {
-                    freeCompilerArgs = freeCompilerArgs + amplifyInternalMarkers.map { "-opt-in=$it" }
+                compilerOptions {
+                    freeCompilerArgs.addAll(amplifyInternalMarkers.map { "-opt-in=$it" })
                 }
             }
         }