+++ build-logic/plugins/src/main/kotlin/ScreenshotConventionPlugin.kt
@@ -1,0 +1,17 @@
+import io.github.takahirom.roborazzi.RoborazziExtension
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.configure
+
+class ScreenshotConventionPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            pluginManager.apply("io.github.takahirom.roborazzi")
+
+            extensions.configure<RoborazziExtension> {
+                outputDir.set(file("src/test/screenshots"))
+            }
+        }
+    }
+}