+++ samples/authenticator/app/src/main/java/com/amplifyframework/ui/sample/authenticator/theme/amplifytheme/Theme.kt
@@ -15,11 +15,16 @@
 
 package com.amplifyframework.ui.sample.authenticator.theme.amplifytheme
 
+import android.app.Activity
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.ViewCompat
 
 private val LightColors = lightColorScheme(
     primary = md_theme_light_primary,
@@ -82,14 +87,19 @@ private val DarkColors = darkColorScheme(
 )
 
 @Composable
-fun AmplifyTheme(
-    useDarkTheme: Boolean = isSystemInDarkTheme(),
-    content: @Composable () -> Unit
-) {
+fun AmplifyTheme(useDarkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
     val colors = if (!useDarkTheme) {
         LightColors
     } else {
         DarkColors
+    }
+
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            (view.context as Activity).window.statusBarColor = colors.primary.toArgb()
+            ViewCompat.getWindowInsetsController(view)?.isAppearanceLightStatusBars = useDarkTheme
+        }
     }
 
     MaterialTheme(